refactor: Moved Edit Period to own component
